\name{Collinearity-package}
\alias{Collinearity-package}
\title{Collinearity}
\description{
Multiple regression is a well-established statistical tool for analysis of complex systems based on many explanatory variables. However, inclusion of many explanatory variables increases the chance of collinearity, which means that one of the explanatory variables is well explained by a linear combination of others. Collinearity is a phenomenon known to deteriorate the model fitness, adversely affecting the inference and the quality of regression parameter estimates.

With this Collinearity package, we provide statistical tools that clarify collinearity issues in multiple regression models. To detect the collinearity of explanatory variables, the function \code{\link{Var_decom_mat}} implements the approach proposed by Belsley (1991) that uses singular-value decomposition of an equilibrated design matrix to compute a matrix of variance decomposition proportions. Whereas the function \code{\link{cofingerprint}} uses bootstrap to assess the stability of estimates in a multiple regression model, the graph generated by the \code{\link{cotograph}} function provides additional insights into the collinearity of explanatory variables. Finally, the function \code{\link{copowerlm}} computes the sample size needed to mitigate the impact of collinearity.
}
\usage{}
\arguments{}
\details{
Package: Collinearity

Type: Package

Version: 1.1.2

Date: 2023-1-10

License: GPL (>=2) 

Depends: R (>= 3.5.0)
}
\value{}
\references{
Belsley, D. A. (1991). \eqn{\textit{Conditioning Diagnostics}} . Probability & Mathematical Statistics S. John
Wiley & Sons, Nashville, TN.
}
\author{Georgios Kazantzidis, Jerome Sepin, Malgorzata Roos

Maintainer: 

Georgios Kazantzidis <George_kazantzidis@outlook.com>

Jerome Sepin <jerome.sepin@gmail.com>
}

