\name{cond_num.lm}
\alias{cond_num.lm}
\title{Condition number}
\description{Calculates the condition number of a matrix}
\usage{
cond_num(object)
}
\arguments{object
  \item{object}{A linear model or a matrix}
   \item{values}{Logical If TRUE, all the eigen value of the matrix are provided} 
}
\value{numeric. The condition number of the object provided. }
\references{Montgomery, Douglas C., Elizabeth A. Peck, and G. Geoffrey Vining. Introduction to linear regression analysis. John Wiley and Sons, 2021.}
\author{Georgios kazantzidis \email{george_kazantzidis@outlook.com}}

