\name{cond_num}
\alias{cond_num}
\title{Condition number}
\description{Calculates the condition number of a linear model design matrix or a matrix}
\usage{
cond_num(object)
}
\arguments{object
  \item{object}{A linear model or a matrix}
}
\value{numeric. The condition number of the object provided. }
\references{Montgomery, Douglas C., Elizabeth A. Peck, and G. Geoffrey Vining. Introduction to linear regression analysis. John Wiley and Sons, 2021.}
\author{Georgios kazantzidis \email{george_kazantzidis@outlook.com}}

