\name{equilibrate_matrix}
\alias{equilibrate_matrix}
\title{equilibrate_matrix function}
\description{
}
\usage{
equilibrate_matrix(object)
}
\arguments{
  \item{object}{
Input is a design matrix X
}
}
\details{}
\value{}
\references{}
\author{Georgios Kazantzidis, Malgorzata Roos}
\note{}
\seealso{}
\examples{}

