\name{n_finder}
\alias{n_finder}
\title{Power calculations for adjustment of collinearity}
\description{Compute the power or determine the sampe size to obtain a target power.}
\usage{n_finder(power=NULL,n = NULL ,alpha=0.05, Delta, sigma, p=3 , voilen=0.5, trouble = 1)}
\arguments{
  \item{power}{power of test (1 minus Type II error probability)}
  \item{n}{number of observations in total}
  \item{alpha}{significance level (Type I error probability)}
  \item{sigma}{experimental error}
  \item{p}{number of parameters in whole model}
  \item{voilen}{variable of interest length}
  \item{trouble}{trouble Diagonal entry of the inverted squared equilibrated design matrix}
}
\details{}
\value{}
\references{}
\author{Jerome Sepin, Malgorzata Roos}

