\name{Var_decom_mat.lm}
\alias{Var_decom_mat.lm}

\title{Var_decom_mat for lm}
\description{Var_decom_mat for lm}
\usage{
Var_decom_mat.lm(object, equilibration)
}
\arguments{
  \item{object}{a linear model.}
  \item{equilibration}{Logical. If TRUE, the design matrix is equilibrated first. Default value is TRUE}
}
\author{Georgios Kazantzidis, Malgorzata Roos}