\name{vdp_svd}
\alias{vdp_svd}
\title{variance-decomposition}
\description{This function uses singular values decomposition to compute variance-decomposition proportions for a design matrix X}
\usage{vdp_svd(object)}
\arguments{
  \item{object}{object is a design matrix X}
}
\details{}
\value{}
\references{}
\author{Georgios Kazantzidis, Malgorzata Roos}

