\name{Collinearity-package}
\alias{Collinearity-package}
\title{Collinearity}
\description{Model choice citeria provide inference regarding the fitness of different models. Collinearity is a phenomenon known to deteriorate the model fitness and causing missleading estimate from the least square estimates approach. With this package, we provide tools to search for collinearity among different models.}
\usage{}
\arguments{}
\details{
Package: Collinearity

Type: Package

Version: 1.1.2

Date: 2023-1-10

License: GPL (>=2) 

Depends: R (>= 3.5.0)
}
\value{}
\references{}
\author{Georgios Kazantzidis, Jerome Sepin, Malgorzata Roos

Maintainer: 

Georgios Kazantzidis <George_kazantzidis@outlook.com>

Jerome Sepin <jerome.sepin@gmail.com>
}

