\name{Var_decom_mat.lm}
\alias{Var_decom_mat.lm}

\title{Var_decom_mat for lm}
\description{ar_decom_mat for lm}
\usage{
Var_decom_mat.lm(object)
}
\arguments{
  \item{object}{a linear model.}
  \item{equilibration}{Logical. If true, the variance decompotision values for the equilibrated matrix are provided.}
}
\author{Georgios Kazantzidis, Malgorzata Roos}