\name{Var_decom_mat}
\alias{Var_decom_mat}
\title{Equilibrated condition indexes}
\description{The variance decompossition matrix along with the equilibrated condition indexes}
\usage{Var_decom_mat(object)}
\arguments{
  \item{object}{a lm object or a matrix}
}
\details{}
\value{}
\references{Belsley, David A. Conditioning diagnostics: Collinearity and weak data in regression.}
\author{Georgios Kazantzidis, Malgorzata Roos, \email{George_kazantzidis@outlook.com}}
\note{}
\seealso{}

